+++ src/__mocks__/context.ts
@@ -96,7 +96,7 @@ export const mockUser = (
 // Generate a mock JWToken
 export const mockToken = async (
   user: User = mockUser(),
-  context?: MyContext
+  context?: MyContext,
 ): Promise<JWTAccessToken> => {
   const email = await user.getEmail(context);
   return {
@@ -129,8 +129,12 @@ export function buildContext(logger: Logger, token: JWTAccessToken = null, cache
 }
 
 // disabling the any since it's the same as above and I think whoever wrote this wanted to avoid the type error
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const buildMockContextWithToken = async (logger: Logger, user: User = mockUser(),  cache: any = null): Promise<MyContext> => {
+export const buildMockContextWithToken = async (
+  logger: Logger,
+  user: User = mockUser(),
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  cache: any = null,
+): Promise<MyContext> => {
   // Only spy on the prototype if user.getEmail is not defined
   if (!user.getEmail && !jest.isMockFunction(User.prototype.getEmail)) {
     jest.spyOn(User.prototype, 'getEmail').mockImplementation(async () => casual.email);