+++ src/resolvers/versionedQuestion.ts
@@ -1,6 +1,7 @@
 import { Resolvers } from "../types";
 import { MyContext } from "../context";
 import { VersionedQuestion } from "../models/VersionedQuestion";
+import { Answer } from "../models/Answer";
 import { AuthenticationError, ForbiddenError, InternalServerError } from "../utils/graphQLErrors";
 import { VersionedQuestionCondition } from "../models/VersionedQuestionCondition";
 import { prepareObjectForLogs } from "../logger";
@@ -8,14 +9,28 @@ import { isAuthorized } from "../services/authService";
 import { GraphQLError } from "graphql";
 import { normaliseDateTime } from "../utils/helpers";
 
+// Define new output structure for the published questions including whether they have an answer
+type VersionedQuestionWithFilled = VersionedQuestion & { hasAnswer: boolean };
+
 export const resolvers: Resolvers = {
   Query: {
-    // return all of the published questions for the specified versioned section
-    publishedQuestions: async (_, { versionedSectionId }, context: MyContext): Promise<VersionedQuestion[]> => {
-      const reference = 'publishedQuestions resolver';
+    // return all published questions for the specified versioned section
+    publishedQuestions: async (_, { planId, versionedSectionId }, context: MyContext): Promise<VersionedQuestionWithFilled[]> => {
+      const reference = 'publishedQuestionsWithAnsweredFlag resolver';
       try {
         if (isAuthorized(context.token)) {
-          return await VersionedQuestion.findByVersionedSectionId(reference, context, versionedSectionId);
+          const questions = await VersionedQuestion.findByVersionedSectionId(reference, context, versionedSectionId);
+
+          // Fetch answers for the questions
+          const questionIds = questions.map(q => q.id);
+          const answers = await Answer.findFilledAnswersByQuestionIds(reference, context, planId, questionIds);
+
+          // Map the answers to the questions
+          const answersMap = new Set(answers.map(a => a.versionedQuestionId));
+          return questions.map(question => ({
+            ...question,
+            hasAnswer: answersMap.has(question.id),
+          })) as VersionedQuestionWithFilled[];
         }
         // Unauthorized!
         throw context?.token ? ForbiddenError() : AuthenticationError();
@@ -45,7 +60,7 @@ export const resolvers: Resolvers = {
   },
 
   VersionedQuestion: {
-    // Chained resolver to return the VersionedQuestionConditionss associated with this VersionedQuestion
+    // Chained resolver to return the VersionedQuestionConditions associated with this VersionedQuestion
     versionedQuestionConditions: async (parent: VersionedQuestion, _, context: MyContext): Promise<VersionedQuestionCondition[]> => {
       return await VersionedQuestionCondition.findByVersionedQuestionId(
         'Chained VersionedQuestion.versionedQuestionConditions',