+++ src/schemas/answer.ts
@@ -8,7 +8,7 @@ export const typeDefs = gql`
     "Get an answer by versionedQuestionId"
     answerByVersionedQuestionId(projectId: Int!, planId: Int!, versionedQuestionId: Int!): Answer
 
-    "Get the sepecific answer"
+    "Get the specific answer"
     answer(projectId: Int!, answerId: Int!): Answer
   }
 
@@ -17,6 +17,12 @@ export const typeDefs = gql`
     addAnswer(planId: Int!, versionedSectionId: Int!, versionedQuestionId: Int!, json: String): Answer
     "Edit an answer"
     updateAnswer(answerId: Int!, json: String): Answer
+    "Add comment for an answer "
+    addAnswerComment(answerId: Int!, commentText: String!): AnswerComment
+    "Update comment for an answer "
+    updateAnswerComment(answerCommentId: Int!, answerId: Int!, commentText: String!): AnswerComment
+    "Remove answer comment"
+    removeAnswerComment(answerCommentId: Int!, answerId: Int!): AnswerComment
   }
 
   "An answer to a question on a Data Managament Plan (DMP)"
@@ -45,6 +51,9 @@ export const typeDefs = gql`
 
     "The comments associated with the answer"
     comments: [AnswerComment!]
+
+    "The feedback comments associated with the answer"
+    feedbackComments: [PlanFeedbackComment!]
   }
 
   "A collection of errors related to the Answer"
@@ -76,6 +85,8 @@ export const typeDefs = gql`
     answerId: Int!
     "The comment"
     commentText: String!
+    "User who made the comment"
+    user: User
   }
 
   "A collection of errors related to the Answer Comment"