+++ CHANGELOG.md
@@ -1,21 +1,53 @@
 # DMP Tool Apollo Server Change Log
 
 ### Added
+- Added a data migration script to update `visibility` to `latestPublishVisibility` in `templates` table. [#405]
+- Added code to update `latestPublishVisibility` whenever a `template` is published [#405]
+- Added model/resolver for `PlanProgress` and nested `progress` in `Plan` resolver [#720]
+- Added checks in `ProjectMember` and `PlanMember` to prevent the deletion of a member if they are the last one [#358]
+- Added a fallback to set a `default` role if none was provided while adding a new `ProjectMember` [#358]
+- Added `findDMPIdsForEmail` helper method to `TokenService` so that the JWT will now contain a list of DMP ids and the user's access level
+- Added `hasPermissionOnPlan` to the `planService` that checks the info contained within the token to determine access instead of making DB calls
+- Added `ProjectFilterOptions` as a possible input for the `myProjects` resolver
+- Added `PlanFeedback` and `PlanFeedbackComments` models [#243]
+- Added `projectCollaboratorCommentsAdded` to `emailService` so that we can email `project collaborators` when new comments added [#243]
+- Added comment mutations to `answers` resolver and created `feedback` resolver [#243]
+- data migration to change collation to `utf8mb4` on all tables
 - added `publishedSection` resolver to `src/resolvers/versionedSection.ts`
 - added `publishedQuestion` resolver to `src/resolvers/versionedQuestion.ts`
 
 ### Updated
+- Updated `answer` and `addAnswerComment` permissions to include a collaborator who has `COMMENT` privileges [#243]
+- Update `removeAnswerComment` and `removeFeedbackComment` to allow deletion by creator, plan creator, or OWN-level collaborator [#243]
+- Created `commentPermissions.ts` service to determine if a comment can be deleted [#243]
+- Updated `docker-compose.yaml` to use the `dmptool-network` which will be shared with other repos
+- Updated `answer`, `fundings`, `members` and `plans` resolver to use new `hasPermissionOnPlan` function
+- Update the `tokenService` to add `dmpIds` array that stores the DMPs the user has access to
+- Updated the `ProjectSearchResult.search` query to provide plan counts by status and filter by status options
+- updated all existing data migrations and scripts to use `utf8mb4` instead of `utf8mb3`
 - updated `PlanSectionProgress` to use better terminology. Changed `sectionId` to `versionedSectionId` (what it really was) and `sectionTtitle` to `title`
 - changed `sections` resolver to `versionedSections` on the `src/resolvers/plan.ts` file and changed the reference for `PlanSearchResult.sections` to `versionedSections`
 
 ### Fixed
+- Fixed bug with maintaining the latest PlanVersion (common standard JSON) when a plan is updated
+- Fixed issue where the title of the DMP in the common standard JSON is using the template title instead of plan title
+- Updated `PlanSectionProgress`model so it correctly shows how many answers have been filled. [#719]
+- Fixed an issue where signup failed because context had been reset to different object
+- Fixed an issue causing the DMP version to not include sections/questions in the narrative if they had not been answered
+- Fixed an issue with the null/undefined check on model queries that use 'searchTerm'
 - When generating a new `versionedTemplate`, we need to deactivate the old ones in the db [#363]
 - Bug with `FunderPopularityResult` in the GraphQL schema that was making `apiTarget` non-nullable
 - added a data migration script to repair bad option based question JSON
 
+### Removed
+- Removed `dmproadmap_template` from the common standard. That information is now stored in `dmproadmap_narrative`
+- Removed `dmpIds` array from the JWT payload.
+- Removed `hasPermissionOnPlan` function from `planService` since JWT no longer has `dmpIds` array
+
 ## v0.2 - Initial deploy to the stage environment
 
 ### Added
+- Added `findFilledAnswersByQuestionIds` which takes plan and question ids and only returns filled answers for those questions.
 - Added data migration to clean up old question JSON so it conforms with new @dmptool/types schemas
 - Added data migration to drop the old `questionTypes` table
 - Added `updatePlanFunding` to allow the update of multiple `planFunding` records [#305]
@@ -239,6 +271,7 @@
 - Old DMPHubAPI datasource and renamed DMPToolAPI to DMPHubAPI since that one had all of the new auth logic
 
 ### Fixed
+- Added `publishedQuestions` to the `versionedQuestion` resolver to return all published questions with boolean for answered or not.
 - Fixed a bug where clients calling `createTemplateVersion` in the `template` resolver would get an error when trying to publish because adding data to `versionedQuestions` required that `questionTypeId` not be null. I added a data-migration script to allow null because I believe we store the question type in the `json` field now and do not require `questionTypeId` [#328]
 - Update profile was not working due to missing `createdById` and `modifiedById` values in db. Added data migration script to populate those fields [#278]
 - Fixed myTemplates query so that `TemplateSearchResult` returns the `ownerDisplayName` specified in schema.